refactor: share chess rules and add camera features
