+++ games/chess3d/README.txt
@@ -6,7 +6,7 @@ What you still need to vendor locally:
 
 1) games/chess3d/lib/three.module.js
 2) games/chess3d/lib/OrbitControls.js (ESM that imports from ./three.module.js)
-3) games/chess3d/engine/chess.min.js (chess.js single-file build)
+3) games/chess/engine/chess.min.js (chess.js single-file build)
 
 After adding those, open games/chess3d/index.html.
 If you see "Engine thinking…" later, you can integrate Stockfish via the AI stubs.
